Arrange RAT questions

Arrange RAT questions

arrange_RAT_questions(rat, by = "original", tRAT_n_offset = 0,
  fixed_number_column = NULL, group_by_column = NULL,
  random_seed = random())

Arguments

rat
Readiness Assessment Test object

by
what to arrange by, options: by="original" leaves the original order as encountered, by="random" generates a random order, by="semi-random" generates a random order except for questions that have the fixed_number_column set, and by="fixed" uses the fixed_number_column (which has to be set for all questions!).

tRAT_n_offset
the offset for the question numbering (if there should be an offset to fit a different part of an IF-AT), only matters if by is not "fixed". Default is 0 (so numbering starts at 1).

fixed_number_column
name of a column in the RAT questions data frame that indicates the fixed question number. Only relevant if by="semi-random" or by="fixed"

group_by_column
name of a column in the RAT questions data frame that indicates which questions to group together (groups will be arranged alphabetically, those with undefined group come last). Only relevant if by="random" or by="semi-random". Throws an error if grouping and any fixed number questions are incompatible (e.g. a question has fixed number 5 but is part of the first group of only 3 questions).

random_seed
can overwrite with a fixed value (e.g. random_seed=42) to get a reproducible "random" order in by="random" or by="semi-random" mode.

Value

returns the RAT with the questions having a new 'number' column