Calculate oxidation blank area/volume

Calculates the oxidation blank area to volume ratio from the direct measurements of the blank peak. For setting this parameter manually, please use set_oxidation_blank.

calculate_oxidation_blank(data, area, volume, criteria = grepl("POR Blank",
  name), update_category = "oxidation blank", quiet = FALSE)

Arguments

data
(can be a grouped_by data set)

area
the area column

volume
the volume column

criteria
the expression used to find the analyses

update_category
value to update the found records category to (pass in NULL if no update of category is desired)

Value

adds the parameter column p.oblank

See also

change_category