Calibrate organic d15 values with the given standards

Calibrate organic d15 values with the given standards

calibrate_d15_org(
  data,
  d15,
  area,
  volume,
  standards = c(`USGS-40` = -4.5, `USGS-41` = 47.6),
  infer_ref_gas = TRUE,
  infer_bgrd = TRUE,
  infer_oblank = TRUE,
  quiet = FALSE
)

Arguments

data
(can be a grouped_by data set)

d15
the d15 column

area
the area (signal) column

standards
a set of isotope standards Note: they are matched to the data by "category" (not by name)

infer_ref_gas
whether to infer reference gas (N2O) isotopic composition from the regression

infer_bgrd
whether to infer bacterial background from the regression

infer_oblank
wheter to infer the organic blank from the regression

Value

introduces the column d15.ocal and parameters p.15o_stds, as well as information on the organic blank A/V ratio p.oblank_ratio and p.oblank_d15 and p.N_blk_amt and bacterial background area p.bgrd_area and p.bgrd_d15

Note

should this also record the actual regression parameters?