Generate RAT multiple choice questions

Generate RAT multiple choice questions

tbl_generate_RAT_choices(rat, iRAT_sel_per_q = 1, iRAT_n_offset = 0,
  answer_layout = "vertical", answer_layout_column = NULL,
  answer_layout_funs = tbl_default_RAT_layouts(), random_seed = random())

Arguments

rat
Readiness Assessment Test object

iRAT_sel_per_q
number of selections per question for the iRAT portion of this test

iRAT_n_offset
number that indicates by how much the iRAT should be offset from the question numbers (e.g. if tRAT starts at a higher IF-AT but iRAT numbers should start low, this could be a negative offset).

answer_layout
how to arrange the answers, layouts supported by default are "vertical" and "horizontal" (recommended for image answers). The layout can be overwritten for individual questions by setting the answer_layout_column parameter. Custom answer layouts can be provided using the answer_layout_funcs parameter.

answer_layout_column
set this parameter to a column name in the questions data frame that has a different layout name for questions that are indended to deviate from the default layout (answer_layout). All layouts must be defined in the answer_layout_funs ("vertical" and "horizontal" by default).

random_seed
can overwrite with a fixed value (e.g. random_seed=42) to get a reproducible "random" order in by="random" or by="semi-random" mode.

answer_layout_funcs
Specific custom answer layouts by providing layout functions that differ from the default. See tbl_default_RAT_layouts for details on how these work.