Convert signal units in continuous flow and dual inlet files

This function can be used to convert the intensity units of a collection of iso_files to a scaled unit (e.g. from mV to V, or nA to mA) or different unit (e.g. from mV to nA or vice-versa). The original signal intensity units are inferred from the naming of the intensity columns. The new units must be a voltage (V) or current (A) but can have any valid SI prefix. Conversion from voltage to current and vice versa requires information about the resistors used in the op amp. This information is automatically retrieved during file read for file formats that contain resistors values (e.g. dxf and did) and are read with the read_method_info=TRUE parameter (see iso_get_resistors_info for details on how to access resistor values in read files). If resistor values are not set in a file, it will not allow automatic conversion between voltage and current. Instead, the R and R_units parameters can be used to provide specific resistor values. However, if R is set, these values will be used for all passed in iso_files.

iso_convert_signals(iso_files, to, R, R_units = NA, quiet = default(quiet))

Arguments

iso_files
collection of iso_file objects

to
what signal unit to convert to

R
resistor value(s). If not specified, will use resitor values from individual iso_files. If specified, must be a named vector with Rm (m=mass) as names (e.g. c(R45=0.3, R46=3)), in units of R_units. If specified, will be used for ALL provided iso_files.

R_units
determined what units resistor values (R) are in, if they are specified. Example R_units = "GOhm" designates that the resistor values provided in R parameter are in Giga-Ohm, i.e. 10^9 Ohm.

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_message_on and iso_turn_info_message_off

Value

the passed in iso_file(s) with changed signal units