Expand file paths

Helper function to expand the provided paths to find data files in folders and subfolders that match any of the specified extensions. Filepaths will be kept as is, only folders will be expanded. Note that this function is rarely called directly. It is used automatically by iso_read_dual_inlet and iso_read_continuous_flow to identify files of interest based on the file paths provided.

iso_expand_paths(path, extensions = c(), root = ".")

Arguments

path

vector of file/folder paths, mixed relative and absolute paths are allowed.

extensions

which extensions to look for? (with or without leading .) - this is typically one or more of the extensions listed by iso_get_supported_file_types

root

root directory for the isofiles. Can be relative to the current working directory (e.g. "data") or an absolute path on the file system (e.g. "/Users/..." or "C:/Data/.."). The default is the current working directory ("."). Can be supplied as a vector of same length as the provided paths if the paths have different roots.

Value

data frame with columns root (root as provided) and path of all the found files.

See also