Filter iso_files

Filter for specific isofiles using file info columns (iso_get_file_info). Works just like dplyr's filter except that it provides the user with some information on what has been filtered. Returns NULL if none of the isofiles' file info matches the filter criteria. You can also use filter directly to filter collections of iso_file objects.

iso_filter_files(iso_files, ..., quiet = default(quiet))

Arguments

iso_files

collection of iso_file objects

...

dplyr-style filter conditions applied based on each file's file_info (see iso_get_file_info)

quiet

whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_messages_on and iso_turn_info_messages_off

See also