Aggregate file info

Combine file information from multiple iso_files. By default all information is included but specific columns can be targeted using the select parameter to select and/or rename columns. File information beyond file_id, file_root, file_path, file_datetime and file_size (in bytes) is only available if the iso_files were read with parameter read_file_info=TRUE.

iso_get_file_info(
  iso_files,
  select = everything(),
  file_specific = FALSE,
  simplify = TRUE,
  quiet = default(quiet)
)

Arguments

iso_files

collection of iso_file objects

select

which columns to select - use c(...) to select multiple, supports all select syntax including renaming columns. File id is always included and cannot be renamed.

file_specific

whether to run the select criteria (...) specifically within each individual file rather than on all files jointly. This is a lot slower but makes it possible to select different columns in different iso_files depending on what exists in each file and is mostly of use when working with data from multiple instruments.

simplify

if set to TRUE (the default), nested value columns in the file info will be unnested as long as they are compatible across file types. Note that file info entries with multiple values still remain nested multi-value (=list) columns even with simplify=TRUE. These can be unnested using unnest.

quiet

whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_messages_on and iso_turn_info_messages_off

Note

this function used to allow selecting/renaming different file_info_columns in different files to the same column. This was a significant speed impediment and only covered very rare use cases. It is still available in the related function iso_select_file_info with a special flag but is no longer the default and not encouraged for use in the frequently called iso_get_file_info.

See also