Retrieve a summary of the problems

Returns a data frame listing how many errors and warnings were encountered for each file. For details on each error/warning, see problems and the iso_problem_functions.

iso_get_problems_summary(
  iso_files,
  problem_files_only = TRUE,
  include_file_info = NULL
)

Arguments

iso_files

collection of iso_file objects

problem_files_only

whether to list only problem files or all files

include_file_info

which file information to include (see iso_get_file_info). Use c(...) to select multiple, supports all select syntax including renaming columns.

Value

data frame with file_id and number of encountered errors and warnings

See also