Example files

The isoreader package comes with a few example files to make it easy to illustrate the functionality.

iso_get_reader_example(filename)

iso_get_reader_examples()

iso_get_reader_examples_folder()

Arguments

filename
the name of the example file for which to retrieve the system path

Details

iso_get_reader_example: retrieve the path to an isoreader example file

iso_get_reader_examples: list of all available isoreader example files

iso_get_reader_examples_folder: path to the location of the reader examples

Examples

iso_get_reader_examples()

#> # A tibble: 12 x 4
#>    filename                 type          software description                  
#>    <chr>                    <chr>         <chr>    <chr>                        
#>  1 continuous_flow_example… continuous f… Isodat   Continuous Flow file format …
#>  2 continuous_flow_example… continuous f… Isodat   Continuous Flow file format …
#>  3 linearity_example.dxf    continuous f… Isodat   Continuous Flow file format …
#>  4 continuous_flow_example… continuous f… ionOS    Continous Flow data archieve 
#>  5 dual_inlet_example.caf   dual inlet    Isodat   Dual Inlet file format (olde…
#>  6 dual_inlet_example.did   dual inlet    Isodat   Dual Inlet file format (newe…
#>  7 dual_inlet_example2.did  dual inlet    Isodat   Dual Inlet file format (newe…
#>  8 dual_inlet_nu_example.t… dual inlet    Nu       Dual Inlet file format       
#>  9 background_scan_example… scan          Isodat   Scan file format             
#> 10 full_scan_example.scn    scan          Isodat   Scan file format             
#> 11 peak_shape_scan_example… scan          Isodat   Scan file format             
#> 12 time_scan_example.scn    scan          Isodat   Scan file format             

iso_get_reader_examples_folder()

#> [1] "/private/var/folders/8z/hxk97xk1763cz5crb3wp7h840000gn/T/RtmpDNXcwZ/temp_libpath2d92fde40f2/isoreader/extdata"