Aggregate vendor computed table data

Aggregate data from the vendor-computed data table. This information is only available if the iso_files were read with parameter read_vendor_data_table=TRUE.

iso_get_vendor_data_table(
  iso_files,
  with_units = FALSE,
  select = everything(),
  include_file_info = NULL,
  with_explicit_units = with_units,
  quiet = default(quiet)
)

Arguments

iso_files

collection of iso_file objects

with_units

this parameter has been DEPRECATED with the introduction of unit-data types (see iso_double_with_units) and will be removed in future versions of isoreader. Please use with_explicit_units instead if you really want columns to have units explicitly in the column name. Alternatively, consider working with the new implicit unit system and convert vendor data tables as needed with iso_make_units_explicit and iso_make_units_implicit.

select

which data columns to select - use c(...) to select multiple, supports all select syntax. By default, all columns are selected.

include_file_info

which file information to include (see iso_get_file_info). Use c(...) to select multiple, supports all select syntax including renaming columns.

with_explicit_units

whether to include units in the column headers of the returned data frame instead of the column data types (see iso_double_with_units). Note that any select conditions have to refer to the column names including the full units.

quiet

whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_messages_on and iso_turn_info_messages_off

See also