Plot chromatogram from continuous flow data

Plot chromatogram from continuous flow data

iso_plot_continuous_flow_data(iso_files, data = c(), time_interval = c(),
  time_interval_units = "seconds", normalize = FALSE, zoom = NULL,
  panel_by = "data", color_by = "file", linetype_by = "none")

Arguments

iso_files
collection of iso_file objects

data
which masses and ratios to plot (e.g. c("44", "45", "45/44")), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using iso_calculate_ratios prior to plotting.

time_interval
which time interval to plot

time_interval_units
which units the time interval is in, default is "seconds"

normalize
whether to normalize all data (default is FALSE, i.e. no normalization). If TRUE, normalizes each trace across all files. Normalizing always scales such that each trace fills the entire height of the plot area. Note that zooming (if zoom is set) is applied after normalizing.

zoom
if not set, automatically scales to the maximum range in the selected time_interval in each panel. If set, scales by the indicated factor, i.e. values > 1 are zoom in, values < 1 are zoom out, baseline always remains the bottom anchor point, zooming is always relative to the max in the plot (even if that is outside visible frame). Note that for overlay plots (panel_bu = "none") zooming is relative to the max in each panel (potentially across different data traces). Also note that zooming only affects masses, ratios are not zoomed.

panel_by
whether to panel data by anything, options are "none" (overlay all), "data" (by mass/ratio data), "file" (panel by files). The default is "data"

color_by
whether to color data by anything, options are the same as for panel_by but the default is "file"

linetype_by
whether to differentiate data by linetype, options are the same as for panel_by but the default is "none". Note that a limited number of linetype_by (6) is defined by default and the plot will fail if a higher number is required unless specified using scale_linetype

See also

Other plot functions: iso_plot_dual_inlet_data, iso_plot_raw_data