Plot mass data from dual inlet files
Source: R/plotting.R

Plot mass data from dual inlet files

iso_plot_dual_inlet_data(iso_files, data = c(), filter = NULL,
  panel = data, color = file_id, linetype = NULL, shape = type, ...)

Arguments

iso_files
collection of iso_file objects

data
which masses and ratios to plot (e.g. c("44", "45", "45/44")), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using iso_calculate_ratios prior to plotting.

filter
any filter condition to apply to the data beyond the masses/ratio selection (param data) and time interval (param time_interval). For details on the available data columns see iso_get_raw_data with parameters gather = TRUE and include_file_info = everything() (i.e. all file info is available for plotting aesthetics).

panel
whether to panel data by anything - any data column is possible (see notes in the filter parameter) but the most commonly used options are panel = NULL (overlay all), panel = data (by mass/ratio data), panel = file_id (panel by files, alternatively use any appropriate file_info column), and panel = type (panel by sample vs standard). Additionally it is possible to panel two variables against each other (i.e. use a facet_grid), e.g. by specifying the formula panel = data ~ file_id (data in the panel rows, files in the panel columns) or panel = data ~ type.The default for this parameter is simple panelling by data.

color
whether to color data by anything, options are the same as for panel but the default is file_id.

linetype
whether to differentiate data by linetype, options are the same as for panel but the default is NULL (i.e. no linetype aesthetic). Note that a limited number of linetypes (6) is defined by default and the plot will fail if a higher number is required unless specified using scale_linetype.

shape
whether to shape data points by anything, options are the same as for panel but the default is type (sample vs standard).

...
deprecated parameters

Note

normalization is not useful for dual inlet data, except potentially between standard and sample - however, for this it is more meaningful to simply plot the relevant ratios together

See also