Map isodat file binary structure.

Map out binary structure for easy visualization (used mostly for error messages and debugging). See the development vignette for details and example application.

map_binary_structure(
  bfile,
  length = 100,
  start = bfile$pos,
  ctrl_blocks = get_ctrl_blocks_config()
)

Arguments

bfile
the binary file, stored in each iso_file under $binary if (and only if) the file was read with iso_turn_debug_on activated before.

length
how many bytes to map

start
at which byte position to start mapping (index 1 based)

ctrl_blocks
named list of block patterns with size, regexp and [optional] replace function