make sure the last plot uses ggplot instead of plotly implementation

make sure the last plot uses ggplot instead of plotly implementation

plot_overview(data, ..., size = NULL, shape = NULL, text = NULL,
  color = category, panel = category)

Arguments

...
the y value columns

size
the data point size aesthetic

shape
the data point shape aesthetic

text
the tooltip text for data points (see make_itext for constructing more complex hover text)

color
what to use for coloring

panel
what to use for panelling (ignored if multiple y columns are supplied)

Note

not sure this is the best way to deal with this, temporary fix while plotly causes trouble Generate a simple overview plot

This functions generates an overview ggplot with either one or multiple y values (latter case makes a tiled plot and ignores the panel parameter). Plot aesthetics like size, text (for hover information if made interactive), color and panelling can be specified.

See also

make_interactive make_itext