Re-read iso_files

Actual multi-purpose file-reread function (not exported) that powers iso_reread_files.

reread_iso_files(
  iso_files,
  ...,
  stop_if_missing = FALSE,
  reread_only_changed_files = FALSE,
  reread_only_outdated_files = FALSE,
  reread_files_without_problems = TRUE,
  reread_files_with_errors = TRUE,
  reread_files_with_warnings = TRUE,
  quiet = default(quiet)
)

Arguments

iso_files

collection of iso_files

...

additional read parameters that should be used for re-reading the iso_files, see iso_read_dual_inlet, iso_read_continuous_flow and iso_read_scan for details (except read_cache which is always set to FALSE to force re-reads).

stop_if_missing

whether to stop re-reading if any of the original data files are missing (if FALSE, will warn about the missing files adding a warning to them, but also re-read those that do exist)

reread_only_changed_files

whether to re-read only files that have since be changed on disc (i.e. have no valid cache file), default FALSE i.e. re-read ALL files

reread_only_outdated_files

whether to re-read only files that were read by an outdated version of isoreader (default FALSE, i.e. re-read ALL files)

reread_files_without_problems

whether to re-read files that had read in without problems the last time (default TRUE)

reread_files_with_errors

whether to re-read files that had read in with errors the last time (default TRUE)

reread_files_with_warnings

whether to re-read files that had read in with warnings the last time (default TRUE)

quiet

whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_messages_on and iso_turn_info_messages_off