Isoreader data structure functions

iso_is_file tests if the object is an iso_file

iso_is_file_list tests if the object is an iso_file list (collection of iso_files)

iso_is_object test if the object is an iso-object (iso_file or iso_file list)

iso_is_dual_inlet tests if an iso_file or iso_file list consists exclusively of dual inlet file objects

iso_is_continuous_flow tests if an iso_file or iso_file list consists exclusively of continuous flow file objects

iso_as_file_list concatenates iso_file and iso_file list object(s) into one combined iso_file list (equivalent to calling c(...)), flattens all passed lists into one list structure, all individual objects and objects within iso_file lists have to be the same type of iso_file, issues warnings if there are duplicate file ids and summarizes all problems in the iso_file list. If duplicates are allowed (discard_duplicates = FALSE), their file IDs will append a #1, #2, #3, etc. to preserve unique file IDs (important for many data aggregation operations).

iso_is_file(x)

iso_is_file_list(x)

iso_is_object(x)

iso_is_dual_inlet(x)

iso_is_continuous_flow(x)

iso_as_file_list(..., discard_duplicates = TRUE)

Arguments

x
an object to test whether it has the specific class

...
iso_file and iso_file_list objects to concatenate

discard_duplicates
whether to automatically discard files with duplicate file IDs (i.e. duplicate file names). If TRUE (the default), only the first files are kept and any files with the same file ID are discarded. If FALSE, all duplicate files are kept but their file IDs are appended with suffix #1, #2, etc.