Aggregate all isofiles data

This function aggregates all isofiles data and returns it in a large data frame with nested columns for each type of information (file_info, raw_data, etc.). For targeted retrieval of specific data iso_get_raw_data, iso_get_file_info, iso_get_vendor_data_table, etc. are much faster and easier to work with. This function is primarily useful for downstream processing pipelines that want to carry all information along. To unnest any of the specific data types (e.g. raw_data), make sure to filter first for the files that have this data type available (e.g. filter(has_raw_data)).

iso_get_data(iso_files, include_file_info = everything(),
  include_raw_data = everything(), include_vendor_data_table = everything(),
  gather = FALSE, with_units = FALSE, with_ratios = FALSE,
  quiet = default(quiet))

Arguments

iso_files
collection of iso_file objects

include_file_info
which file information to include (see iso_get_file_info). Use c(...) to select multiple, supports all select syntax including renaming columns.

include_raw_data
which columns from the raw data to include use c(...) to select multiple, supports all select syntax including renaming columns. Includes all columns by default.

include_vendor_data_table
which columns from the vendor data table to include - use c(...) to select multiple, supports all select syntax including renaming columns. Includes all columns by default.

gather
whether to gather raw data into long format (e.g. for ease of use in plotting). Not that the select parameter applies to the data columns BEFORE gathering.

with_units
whether to include units in the column headers (if there are any) or not (default is FALSE). Note that this can slow down the function significantely.

with_ratios
whether to include ratios or just standard delta values

quiet
whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to iso_turn_info_message_on and iso_turn_info_message_off

Value

data_frame with file_ids, file_types and nested data frames for each data type (file_info, raw_data, vendor_data_table, etc.)

See also