Example files

The isoreader package comes with a few example files to make it easy to illustrate the functionality.

iso_get_reader_example(filename)

iso_get_reader_examples()

iso_get_reader_examples_folder()

Arguments

filename

the name of the example file for which to retrieve the system path

Details

iso_get_reader_example: retrieve the path to an isoreader example file

iso_get_reader_examples: list of all available isoreader example files

iso_get_reader_examples_folder: path to the location of the reader examples

Examples

iso_get_reader_examples()
#> # A tibble: 10 × 4
#>    filename                     type            software description            
#>    <chr>                        <chr>           <chr>    <chr>                  
#>  1 continuous_flow_example.cf   continuous flow Isodat   Continuous Flow file f…
#>  2 continuous_flow_example.dxf  continuous flow Isodat   Continuous Flow file f…
#>  3 continuous_flow_example.iarc continuous flow ionOS    Continuous Flow data a…
#>  4 dual_inlet_example.caf       dual inlet      Isodat   Dual Inlet file format…
#>  5 dual_inlet_example.did       dual inlet      Isodat   Dual Inlet file format…
#>  6 dual_inlet_nu_example.txt    dual inlet      Nu       Dual Inlet file format 
#>  7 background_scan_example.scn  scan            Isodat   Scan file format       
#>  8 full_scan_example.scn        scan            Isodat   Scan file format       
#>  9 peak_shape_scan_example.scn  scan            Isodat   Scan file format       
#> 10 time_scan_example.scn        scan            Isodat   Scan file format       
iso_get_reader_examples_folder()
#> [1] "/home/runner/work/_temp/Library/isoreader/extdata"