Plot chromatogram from continuous flow data

Plot chromatogram from continuous flow data

iso_plot_continuous_flow_data(iso_files, data = c(), time_interval = c(),
  time_interval_units = "seconds", filter = NULL, normalize = FALSE,
  zoom = NULL, panel = data, color = file_id, linetype = NULL,
  label = file_id, ...)

Arguments

iso_files
collection of iso_file objects

data
which masses and ratios to plot (e.g. c("44", "45", "45/44")), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using iso_calculate_ratios prior to plotting.

time_interval
which time interval to plot

time_interval_units
which units the time interval is in, default is "seconds"

filter
any filter condition to apply to the data beyond the masses/ratio selection (param data) and time interval (param time_interval). For details on the available data columns see iso_get_raw_data with parameters gather = TRUE and include_file_info = everything() (i.e. all file info is available for plotting aesthetics).

normalize
whether to normalize all data (default is FALSE, i.e. no normalization). If TRUE, normalizes each trace across all files. Normalizing always scales such that each trace fills the entire height of the plot area. Note that zooming (if zoom is set) is applied after normalizing.

zoom
if not set, automatically scales to the maximum range in the selected time_interval in each panel. If set, scales by the indicated factor, i.e. values > 1 are zoom in, values < 1 are zoom out, baseline always remains the bottom anchor point, zooming is always relative to the max in the plot (even if that is outside visible frame). Note that for overlay plots (panel_bu = "none") zooming is relative to the max in each panel (potentially across different data traces). Also note that zooming only affects masses, ratios are not zoomed.

panel
whether to panel data by anything - any data column is possible (see notes in the filter parameter) but the most commonly used options are panel = NULL (overlay all), panel = data (by mass/ratio data), panel = file_id (panel by files, alternatively use any appropriate file_info column). The default is panelling by data.

color
whether to color data by anything, options are the same as for panel but the default is file_id.

linetype
whether to differentiate data by linetype, options are the same as for panel but the default is NULL (i.e. no linetype aesthetic). Note that a limited number of linetypes (6) is defined by default and the plot will fail if a higher number is required unless specified using scale_linetype.

label
this is primarily of use for turning these into interactive plots via ggplotly as it present as an additional mousover label. Any unique file identifier is a useful choice, the default is file_id.

...
deprecated parameters

See also